% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empfib_GF}
\alias{empfib_GF}
\title{GenomicFiles instance for 20 GTEx Lung samples (with notations on emphysema or fibrosis)}
\format{
GenomicFiles
}
\usage{
empfib_GF
}
\description{
GenomicFiles instance for 20 GTEx Lung samples (with notations on emphysema or fibrosis)
}
\keyword{datasets}
