% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expt.R
\name{expt}
\alias{expt}
\title{helper function for MAE filtering, just returns experiments(mae)[[tag]]}
\usage{
expt(.data, tag)
}
\arguments{
\item{.data}{MultiAssayExperiment-like entity}

\item{tag}{character(1) should name an experiment in MAE}
}
\description{
helper function for MAE filtering, just returns experiments(mae)[[tag]]
}
