% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.DFrame}
\alias{filter.DFrame}
\title{S3 support for inline filter}
\usage{
\method{filter}{DFrame}(.data, ...)
}
\arguments{
\item{.data}{DataFrame instance}

\item{\dots}{as needed}
}
\description{
S3 support for inline filter
}
