% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr38s}
\alias{gr38s}
\title{GRanges with gene addresses from TxDb hg38}
\format{
GRanges
}
\usage{
gr38s
}
\description{
GRanges with gene addresses from TxDb hg38
}
\examples{
data(gr38s)
gr38s
}
\keyword{datasets}
