% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTExMAE}
\alias{GTExMAE}
\title{PFB extract of GTEx converted to MultiAssayExperiment}
\format{
MultiAssayExperiment
}
\usage{
GTExMAE
}
\description{
PFB extract of GTEx converted to MultiAssayExperiment
}
\examples{
data(GTExMAE)
GTExMAE
}
\keyword{datasets}
