% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_names.R
\name{simplify_names}
\alias{simplify_names}
\title{simplify names of PFB export}
\usage{
simplify_names(.data)
}
\arguments{
\item{.data}{input data.frame-like entity}
}
\description{
simplify names of PFB export
}
