% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntplot.R
\name{tntplot}
\alias{tntplot}
\title{Use TnT to visualize a GRanges segment, assumes an mcols element 'score' is present, numeric-like}
\usage{
tntplot(
  gr,
  scorecolor = "lightblue",
  genecolor = "gold",
  gt = NULL,
  viewradius = 1e+05,
  coordradius = 2e+05
)
}
\arguments{
\item{gr}{GRanges instance}

\item{scorecolor}{character(1) color used on PinTrack for scores}

\item{genecolor}{character(1) color used on GeneTrackFromTxDb for gene regions}

\item{gt}{optional GeneTrack to avoid recomputing; defaults to NULL}

\item{viewradius}{numeric(1) radius around given `gr` for display}

\item{coordradius}{numeric(1) radius around region to allow panning}
}
\description{
Use TnT to visualize a GRanges segment, assumes an mcols element 'score' is present, numeric-like
}
