% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_emt}
\alias{msig_emt}
\title{genes in MSigDb HALLMARK_EPITHELIAL_MESENCHYMAL_TRANSITION}
\format{
character()
}
\usage{
msig_emt
}
\description{
genes in MSigDb HALLMARK_EPITHELIAL_MESENCHYMAL_TRANSITION
}
\examples{
data(msig_emt)
head(msig_emt)
}
\keyword{datasets}
